+++ src/cluster/BridgedDeviceBasicInformationCluster.ts
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-empty-object-type */
 /**
  * @license
  * Copyright 2022-2024 Matter.js Authors
@@ -58,7 +59,7 @@ export namespace BridgedDeviceBasicInformation {
     finish: TlvField(0, TlvEnum<ProductFinish>()),
     primaryColor: TlvField(1, TlvNullable(TlvEnum<Color>())),
   });
-  // eslint-disable-next-line @typescript-eslint/no-empty-interface
+
   export interface ProductAppearanceStruct extends TypeFromSchema<typeof TlvProductAppearanceStruct> {}
 
   /**
@@ -73,7 +74,7 @@ export namespace BridgedDeviceBasicInformation {
    *
    * @see {@link MatterSpecification.v11.Core} § 9.13.5
    */
-  // eslint-disable-next-line @typescript-eslint/no-empty-interface
+
   export interface StartUpEvent extends TypeFromSchema<typeof TlvStartUpEvent> {}
 
   /**
@@ -88,7 +89,7 @@ export namespace BridgedDeviceBasicInformation {
    *
    * @see {@link MatterSpecification.v11.Core} § 9.13.5.1
    */
-  // eslint-disable-next-line @typescript-eslint/no-empty-interface
+
   export interface ReachableChangedEvent extends TypeFromSchema<typeof TlvReachableChangedEvent> {}
 
   /**
@@ -243,7 +244,7 @@ export namespace BridgedDeviceBasicInformation {
    *
    * @see {@link MatterSpecification.v11.Core} § 9.13
    */
-  // eslint-disable-next-line @typescript-eslint/no-empty-interface
+
   export interface Cluster extends Identity<typeof ClusterInstance> {}
 
   export const Cluster: Cluster = ClusterInstance;